Use ES6 features in code and exttact HTTP API code to a separate file
